pyproject.toml
[tool.poetry]
name = "bootstrap-fastapi-service"
version = "0.1.0"
description = "Bootstrap FastAPI backend with src layout, dotenv configuration, scheduler, and logging."
authors = ["Bootstrapper <dev@example.com>"]
readme = "README.md"
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
tzdata = "^2024.1"
httpx = "^0.27.2"
websockets = "^12.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.0"
pytest = "^8.1.0"
pytest-asyncio = "^0.23.0"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
